+++ src/features/order/components/OrderFromCustomer.jsx
@@ -1,13 +1,14 @@
 import { useState, useEffect, useCallback } from "react";
 import { useNavigate } from "react-router-dom"; // import useNavigate
-import CommonButton from "../components/CommonButton";
-import { IconArrow } from "../icons/IconArrow";
-import useSocket from "../hooks/socketIoHook";
+
+import { IconArrow } from "../../../icons/IconArrow";
+import useSocket from "../../../hooks/socketIoHook";
 import { toast } from "react-toastify";
-import useAuth from "../hooks/authHook";
-import ModalCommon from "./ModalCommon";
+import useAuth from "../../../hooks/authHook";
+import ModalCommon from "../../../components/ModalCommon";
 import { motion, AnimatePresence } from "framer-motion";
-import { IconLogoHop } from "../icons/IconLogoHop";
+import { IconLogoHop } from "../../../icons/IconLogoHop";
+import CommonButton from "../../../components/CommonButton";
 
 let isAccept = false;
 export default function OrderFromCustomer() {
